phase2_solution/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
openai>=1.3.7
python-dotenv>=1.0.0
pydantic>=1.10.0